Library/requirements.txt
asgiref==3.5.0
Django==3.0.7
djangorestframework==3.13.1
psycopg2==2.8.6
pytz==2021.3
sqlparse==0.4.2
